+Wrote logic for refreshing data of equip menu
I said this would take a lot of work and it did... 1780 lines of code later and I'm at the brink of insanity.  It updates the information properly, I just have to add the system that allows you to equip weapons armor and trinket to test if the refreshes properly when equipping items.