Worked on status effects more
Majority of effects should work, will need to do more with those that effect Sanity, Pow, and Will.
Need to fix bug with vomiting not doing damage.
EnemyTarget should be used to select targets for enemy abilities
Added code so that enemies should select ability to match the row the target is in.
Will work on swap abilities tomorrow